requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
boto3==1.26.157
botocore==1.29.157
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
deepgram-sdk==2.5.0
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
multidict==6.0.4
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
websockets==11.0.3
Werkzeug==2.3.6
yarl==1.9.2
